Reduce unistd
git-svn-id: file:///dev/shm/somefilter.svn@698 e102df66-1e2e-11dd-9b44-c24451a4db5e

+++ util/file.hh
@@ -1,8 +1,10 @@
 #ifndef UTIL_FILE__
 #define UTIL_FILE__
 
+#include <cstddef>
 #include <cstdio>
-#include <unistd.h>
+
+#include <inttypes.h>
 
 namespace util {
 
@@ -61,8 +63,8 @@ int OpenReadOrThrow(const char *name);
 int CreateOrThrow(const char *name);
 
 // Return value for SizeFile when it can't size properly.  
-const off_t kBadSize = -1;
-off_t SizeFile(int fd);
+const uint64_t kBadSize = (uint64_t)-1;
+uint64_t SizeFile(int fd);
 
 void ReadOrThrow(int fd, void *to, std::size_t size);
 std::size_t ReadOrEOF(int fd, void *to_void, std::size_t amount);
